querydsl 예제

+++ src/main/java/model/Employee.java
@@ -51,7 +51,7 @@ public class Employee {
     private Integer mgrid;
 
     @Column(name = "DEPARTMENT_ID")
-    private Integer deptid;
+    private Long deptid;
 
     @ManyToOne   // 테이블 연관 관계 = 다 : 1
     @JoinColumn(name="department_id")
